+++ examples/generate-dts/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @farmfe-examples/generate-dts
 
+## 1.0.10
+
+### Patch Changes
+
+- Updated dependencies [dcce5f6]
+  - @farmfe/js-plugin-dts@0.2.0
+
+## 1.0.9
+
+### Patch Changes
+
+- Updated dependencies [0f93f94]
+  - @farmfe/core@0.13.10
+  - @farmfe/js-plugin-dts@0.1.2
+  - @farmfe/cli@0.5.8
+
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies [62e6630]
+  - @farmfe/core@0.13.9
+  - @farmfe/js-plugin-dts@0.1.2
+  - @farmfe/cli@0.5.8
+
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies [6a73829]
+  - @farmfe/core@0.13.8
+  - @farmfe/js-plugin-dts@0.1.2
+  - @farmfe/cli@0.5.8
+
 ## 1.0.6
 
 ### Patch Changes