+++ examples/script-entry/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @farmfe-examples/script-entry
 
+## 0.0.75
+
+### Patch Changes
+
+- Updated dependencies [0f93f94]
+  - @farmfe/core@0.13.10
+  - @farmfe/cli@0.5.8
+
+## 0.0.74
+
+### Patch Changes
+
+- Updated dependencies [62e6630]
+  - @farmfe/core@0.13.9
+  - @farmfe/cli@0.5.8
+
+## 0.0.73
+
+### Patch Changes
+
+- Updated dependencies [6a73829]
+  - @farmfe/core@0.13.8
+  - @farmfe/cli@0.5.8
+
 ## 0.0.72
 
 ### Patch Changes